Safari-611.1.10/LayoutTests/crypto/subtle/ecdsa-import-jwk-public-key-alg-384-expected.txt

Test importing a JWK ECDSA public key with ES384

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDSA'
PASS publicKey.algorithm.namedCurve is 'P-384'
PASS publicKey.usages is ['verify']
PASS successfullyParsed is true

TEST COMPLETE
